Decouple `InMemoryModuleArtifactsCache` from `DefaultModuleArtifactsCache`
